+++ config.sample.js
@@ -26,5 +26,5 @@ module.exports = {
   DEBUG_SQL: false, // Traces all SQL calls
   DEBUG_REFRESH_BROWSER: true, // Refreshes the browser automatically upon code changes
   DEBUG_ADMIN: false, // Lets you access the "/admin" URL even when logged out
-  DEBUG_ALLOW_INVALID_INVITE_KEYS: false // Lets you register without an invite key,
+  DEBUG_ENABLE_INVITE_SYSTEM: false // Enables the now abandoned Steam-like key system for creating accounts
 }